requirements.txt
flask
openai
azure-ai-formrecognizer
